Open snap automatically and misc.
Remove the start programming button and have snap load automatically when a robot is connected. Add a modal popup for the case when the app is loaded in an unsupported browser, and for the case where the computer does not have bluetooth. Add other snap projects to cache. Remove unneeded code from hummingbird snap project.