Name update and more brython fixes
Change the name of the apps to "BirdBrain Brython" and "BirdBrain Snap!". Brython: Change device orientation text so that the second word isn't capitalized to match Python. Remove unused code in BirdBrain module. Add error message to getButton. Issue stopall command to robot when first seen in brython. This allows for the brython iframe to be reloaded to stop a script. Make more adjustments to the wrapPython script. Keep the run button from acting if there is already a script running. Add stop button which will reload the iframe to stop the currently running script.